tweak(tupaiaWeb): RN-249: getting x-Name in viz export (#5886)
* first attempt

* merge: dev -> rn-249 (#5895)

* tweak(centralServer): RN-1351: Handle bad DHIS2 sync requests (#5887)

Throw an error and register bad requests on dhis sync queue

* tweak(datatrakWeb): RN-1438: Upgrade React Query to V4 (#5868)

* Bumping react-query

* Update render.tsx

* fixing Reports.test.tsx

* navigation bug fix

* removed logging

* tweak(adminPanel): RN-1439: Upgrade React Query to V4 (#5875)

* bumping to v4

* Update yarn.lock

* tweak(psss): RN-1442: Upgrade React Query to V4 (#5874)

* upgrade

* Update yarn.lock

* tweak(lesmis): RN-1441: Upgrade React Query to V4 (#5873)

* upgrade

* Update yarn.lock

* tweak(tupaiaWeb): RN-1440: Upgrade React Query to V4 (#5870)

* Bumping react-query

* Bumping react-query to v4

* Update render.tsx

* GitAction fix

* fixing Reports.test.tsx

* navigation bug fix

* review update

* removed logging

* Update yarn.lock

* Update yarn.lock

* Update VerifyEmailPage.tsx

* Update useReportPreview.js

* fix(adminPanel): Miscellaneous Bug fixes (Tiny) (#5891)

* bugfix

* update

* updates

* tweak(tupaiaWeb): RN-1437: Download files visual restyle (#5889)

* Common download files component

* PR fix

---------


* refactor legend

---------
