+++ packages/ui-chart-components/src/components/Legend/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
+ */
+
+export { getCartesianChartLegend } from './getCartesianChartLegend';
+export { getPieChartLegend } from './getPieChartLegend';
