+++ packages/ui-chart-components/src/components/Reference/index.ts
@@ -4,7 +4,6 @@
  */
 
 export { ChartTooltip } from './ChartTooltip';
-export * from './Legend';
 export { ReferenceLabel } from './ReferenceLabel';
 export { ReferenceLines } from './ReferenceLines';
 export { TooltipContainer } from './TooltipContainer';
